fix: new KYC test; add gas limit for revokeKyc

+++ contracts/contracts/extensions/Interfaces/ITokenOwner.sol
@@ -21,7 +21,7 @@ interface ITokenOwner {
      *
      * @param code The code to check
      */
-    error ResponseCodeInvalid(int256 code);
+    error ResponseCodeInvalid(int64 code);
 
     /**
      * @dev Emitted when the provided `value` is not greater than `ref`
